Support for MP decomposition (books) in SQLite format

See merge request fuw_software/svarog2!62

+++ svarog/src/main/java/org/signalml/app/view/book/BookPlot.java
@@ -1435,12 +1435,12 @@ public class BookPlot extends JComponent implements PropertyChangeListener {
 			return null;
 		}
 
-		int frequency = atom.getNaturalFrequency();
+		float frequency = atom.getNaturalFrequency();
 		if (frequency < naturalMinFrequency || frequency > naturalMaxFrequency) {
 			return null;
 		}
 
-		int position = atom.getPosition();
+		float position = atom.getPosition();
 		if (position < pointMinPosition || position > pointMaxPosition) {
 			return null;
 		}
